Do not use index for filter-branch directory-filter
Directory-filter option of git filter-branch doesn't require the tree actually loaded into index.
Skip on it when directory-filter is specified alone (without tree-filter or index-filter).
The only point of doing that would be if "git write-tree" would recompute trees affected by object replacement,
but, unfortunately, write-tree doesn't do that.
This optimization reduced time spent by 40% when run on an SSD.