filter-branch: Add tree rebuild option: --diff-tree-filter, and --index-pre-filter
This commit adds two command-line options:

--diff-tree-filter <command> - filter to modify file path or mode (for example, to make sure
all scripts are executable). You can also rename paths in this filter, and drop files.
Diff-tree filter gets called for each file modified by a commit,
including those not changed by EOL normalization.

Diff-tree filter takes a single line from stdin, formatted as:
<mode><SP><SHA1><SP>0<TAB><pathname>

and can output a modified line formatted the same way.
If the filter doesn't output the line, the file gets dropped.
If some commits of your repository already submit a .gitattributes file, it's advised
that the diff-tree filter detects a possible overwrite of .gitattributes,
and substitutes the file hash with the same you want added in pre-normalize.

As a side effect of diff-tree filter, the trees are rebuilt with the renamed objects substituted to the replacement hash.
This means, the result of "filter-branch --diff-tree-filter" will not need the renames carried on.
Simple "filter-branch --directory-filter/tree-filter/index-filter" doesn't result in the replacement substitution,
which means the replacement refs still need to be carried with the rebuilt repository.

If you only want trees rebuilt without any custom adjustments (to only apply all replacements),
use --diff-tree-filter cat

This commit adds a second option --index-pre-filter, which, even though it was designed to assist --diff-tree-filter usage,
can also be also used by itself.

--index-pre-filter <command> - is invoked before the changes between previous and the current
commit are applied to the index, or before --tree-filter. Regular --index-filter is invoked
after --tree-filter, which is not always what we want.

You can use this filter is to add .gitattributes blob to the index, usually at the root of the work tree.
Use "git update-index" command for that.
Since it's usually the same blob for all commits, pre-made it by using "git hash-object",
and feed its SHA1 to update-index.